python camera/requirements.txt
opencv-python
numpy==1.19.4
Pillow
pyserial